ce-auth-proxy/requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
PyJWT==2.0.1
requests==2.25.1
trompace-client==0.2
typing-extensions==3.7.4.3
urllib3==1.26.3
websockets==8.1
Werkzeug==1.0.1
yarl==1.6.3
